matrix-vector product, move matrix to its own header
